% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicomp.R
\name{bicomp}
\alias{bicomp}
\title{Bilinear Decomposition of a Matrix}
\usage{
bicomp(x, delta = c(1, 1, 1, 1), which = 0L:4L)
}
\arguments{
\item{x}{A matrix to be decomposed.}

\item{delta}{A vector of length four with 0/1 entries which controls the type of
decomposition made.}

\item{which}{A vector giving the elements to return, with 0 = original data, 1 = overall means, 
2 = row means, 3 = column means and 4 = interactions.}
}
\value{
An object of class \code{bicomp}, possible also inheriting from class \code{data.frame}, 
which is either a named list with the required components, or a single matrix if a single 
component is requested. An additional attribute \code{return_type} gives information on the type 
of matrices returned.
}
\description{
Decomposes a matrix into an overall mean matrix, row margins matrix, column margins matrix
and an interaction matrix, depending on \code{delta}.
}
