% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carray.R
\name{carray}
\alias{carray}
\title{Double-Centre a Three-way Array}
\usage{
carray(array, margin = 3L, rows = TRUE, columns = TRUE)
}
\arguments{
\item{array}{A three-way array}

\item{margin}{The way of the array over which the centring must be done}

\item{rows}{Logical indicating whether to centre the rows of the matrix slices}

\item{columns}{Logical indicating whether to centre the columns of the matrix slices}
}
\description{
Double-centre the matrix slices of a three-way array.
}
