% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsim.R
\name{cfsim}
\alias{cfsim}
\title{Compare Simulation Results}
\usage{
cfsim(fitted, actual, method = c("diag", "cRand"))
}
\arguments{
\item{fitted}{An object of class \code{lsbclust} containing the fitted results.}

\item{actual}{An object of class \code{lsbclust_sim} containing the simulated data.}

\item{method}{The type of statistics to calculate, passed to \code{\link{cl_agreement}}}
}
\description{
Generic function to compare simulation results in \pkg{lsbclust}.
}
\seealso{
\code{\link{cfsim.lsbclust}}, \code{\link{cfsim.T3Clusf}}
}
