% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClustMeans}
\alias{ClustMeans}
\title{C++ Function for Cluster Means}
\usage{
ClustMeans(nclust, start, data)
}
\arguments{
\item{nclust}{The number of clusters.}

\item{start}{The current clustering vector.}

\item{data}{The concatenated data, with J * K rows and N columns}
}
\value{
A numeric matrix with \code{nclust} rows and \code{J*K} columns.
}
\description{
This function calculates the cluster means in vectorized form based on the current
value of the clustering vector.
}
