% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.lsbclust.R
\name{cmat}
\alias{cmat}
\title{Centring Matrix}
\usage{
cmat(k)
}
\arguments{
\item{k}{An integer determining the dimensions of the centring matrix.}
}
\description{
A utility function for calculating centring matrices.
}
