% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.akmeans.R
\name{fitted.akmeans}
\alias{fitted.akmeans}
\title{Extract Fitted Values for akmeans}
\usage{
\method{fitted}{akmeans}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"akmeans"}}

\item{\dots}{Unimplemented}
}
\value{
An array approximating the original data
}
\description{
An S3 method for \code{\link{fitted}} for class \code{"akmeans"}.
}
\seealso{
\code{\link{akmeans}}
}
