% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.lsbclust.R
\name{fitted.lsbclust}
\alias{fitted.lsbclust}
\title{Extract Fitted Values for LSBCLUST}
\usage{
\method{fitted}{lsbclust}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"lsbclust"}}

\item{\dots}{Unimplemented}
}
\value{
An array approximating the original data
}
\description{
An S3 method for \code{\link{fitted}} for class \code{"lsbclust"}.
}
\seealso{
\code{\link{lsbclust}}
}
