% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.T3Clusf.R
\name{fitted.T3Clusf}
\alias{fitted.T3Clusf}
\title{Extract Fitted Values for T3Clusf}
\usage{
\method{fitted}{T3Clusf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"T3Clusf"}}

\item{\dots}{Unimplemented}
}
\value{
An array approximating the original data
}
\description{
An S3 method for \code{\link{fitted}} for class \code{"T3Clusf"}.
}
\seealso{
\code{\link{T3Clusf}}
}
