% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsbclust-package.R
\docType{package}
\name{lsbclust-package}
\alias{lsbclust-package}
\title{Least Squares Latent Class Matrix Factorization}
\description{
Funtions for least squares latent class matrix factorizations.
}
\references{
Van Rosmalen, J., Van Herk, H., & Groenen, P. J. F. (2010). Identifying response styles: A latent-class bilinear multinomial logit model. \emph{Journal of Marketing Research}, 47(1), 157-172.
}
\author{
Pieter C. Schoonees [aut, cre], Patrick J.F. Groenen [aut]
}
\keyword{package}
