% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cluedefinitions.R
\name{cl_class_ids.int.lsbclust}
\alias{cl_class_ids.int.lsbclust}
\alias{is.cl_hard_partition.int.lsbclust}
\alias{is.cl_partition.int.lsbclust}
\title{S3 Methods for Integration into \pkg{clue} Framework}
\usage{
\method{cl_class_ids}{int.lsbclust}(x)

\method{is.cl_partition}{int.lsbclust}(x)

\method{is.cl_hard_partition}{int.lsbclust}(x)
}
\arguments{
\item{x}{An object of class \code{int.lsclust}}
}
\description{
These methods integrates the class \code{int.lsbclust} into the framework set out in
package \pkg{clue}.
}

