% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanbiplot.R
\name{meanbiplot}
\alias{meanbiplot}
\title{Biplots of}
\usage{
meanbiplot(rows, cols)
}
\arguments{
\item{rows}{A list of matrices representing the rows}

\item{cols}{A list of matrices representing the columns}
}
\description{
Construct simple two-dimensional biplots given matrices representing the rows
and columns of a two-dimensional matrix using \pkg{ggplot2}.
}
\examples{
set.seed(1)
dat <- rlsbclust(ndata = 1, nobs = 100, size = c(10, 8), nclust = c(5, 4, 6, 5))
meanbiplot(dat[[1]]$interactions$C, dat[[1]]$interactions$D)
}
