% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bicomp.R
\name{plot.bicomp}
\alias{plot.bicomp}
\title{Plot a \code{bicomp} Object}
\usage{
\method{plot}{bicomp}(x, which = 0L:4L, arrange = TRUE,
  col = c("red4", "beige", "blue4"), strip.legend = TRUE,
  add.titles = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{bicomp}.}

\item{which}{A numeric vector indicating which matrices to plot, with 0 = original data, 
1 = overall means, 2 = row means, 3 = column means and 4 = interactions.}

\item{arrange}{Logical indicating whether the arrange  the plots side-by-side
via \code{\link{grid.arrange}} or not.}

\item{col}{A character vector of length three giving the parameters 
\code{low}, \code{mid} and \code{high} for \code{\link{scale_fill_gradient2}}.}

\item{strip.legend}{Logical indicating whether to strip the legend off the plot or not.}

\item{add.titles}{Logical indicating whether to add titles to the plots or not.}

\item{\dots}{Additional arguments to \code{\link{theme}}.}
}
\description{
Plot method for an object of class \code{bicomp} (see \code{\link{bicomp}}).
}
