% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.col.kmeans.R
\name{plot.col.kmeans}
\alias{plot.col.kmeans}
\title{Plot method for class 'col.kmeans'}
\usage{
\method{plot}{col.kmeans}(x, which = 1L, ...)
}
\arguments{
\item{x}{An object of class \code{col.kmeans}}

\item{which}{Which type of plot to produce (only 3 types are implemented).}

\item{\dots}{additional arguments passed to \code{\link{theme}}.}
}
\description{
Simple plot method for object of class 'col.kmeans' as output by \code{\link{orc.lsbclust}}.
}
\examples{
data("dcars")
m <- orc.lsbclust(data = dcars, margin = 3, delta = c(1,1,1,1), nclust = 5, type = "columns")
plot(m)
}
\author{
Pieter C. Schoonees
}
\keyword{hplot}
