% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ovl.kmeans.R
\name{plot.ovl.kmeans}
\alias{plot.ovl.kmeans}
\title{Plot method for class 'ovl.kmeans'}
\usage{
\method{plot}{ovl.kmeans}(x, which = 1L, ...)
}
\arguments{
\item{x}{An object of class \code{ovl.kmeans}}

\item{which}{Which type of plot to produce. Currently only \code{which = 1} is implemented.}

\item{\dots}{additional arguments passed to \code{\link{theme}}.}
}
\description{
Simple plot method for object of class 'ovl.kmeans' as output by \code{\link{orc.lsbclust}}.
}
\examples{
data("dcars")
m <- orc.lsbclust(data = dcars, margin = 3, delta = c(1,1,1,1), nclust = 5, type = "overall")
plot(m)
}
\author{
Pieter C. Schoonees
}
\keyword{hplot}
