% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.step.lsbclust.R
\name{plot.step.lsbclust}
\alias{plot.step.lsbclust}
\title{Plot method for class 'step.lsbclust'}
\usage{
\method{plot}{step.lsbclust}(x, which = 1L:5L, col.all = NULL,
  arrange = FALSE, chull = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{step.lsbclust}}

\item{which}{Which type of plot to produce.}

\item{col.all}{A character vector of length one indicating which of \code{"overall"}, 
\code{"rows"}, \code{"columns"} or \code{"interactions"} should be mapped to colour in the
plot for all possible models. Care needs to be taken that the stated component is included in
the fit.}

\item{arrange}{Logical indicating whether the arrange  the plots side-by-side
via \code{\link{grid.arrange}} or not.}

\item{chull}{Logical indicating whether to plot the estimated convex hull or not.}

\item{\dots}{additional arguments passed to \code{\link{theme}}.}
}
\description{
Plot 'step.lsbclust' objects.
}
\author{
Pieter C. Schoonees
}
\keyword{hplot}
