% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.T3Clusf.R
\name{plot.T3Clusf}
\alias{plot.T3Clusf}
\title{Plot Method for Class 'T3Clusf'}
\usage{
\method{plot}{T3Clusf}(x, which = seq_len(nclust), arrange = FALSE,
  ...)
}
\arguments{
\item{x}{An object of class \code{T3Clusf}.}

\item{which}{An integer vector indicating which item segments to plot.}

\item{arrange}{Logical indicating whether to arrange the plots on a single page or not}

\item{\dots}{Additional arguments to \code{\link{theme}}}
}
\description{
Two-dimensional plot method for object of class 'T3Clusf' as output by \code{\link{T3Clusf}}.
}
\keyword{hplot}
