% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lsbclust.R
\name{print.lsbclust}
\alias{print.lsbclust}
\title{Print method for object of class 'lsbclust'}
\usage{
\method{print}{lsbclust}(x, ...)
}
\arguments{
\item{x}{An object of class 'lsbclust'}

\item{\dots}{Unimplemented.}
}
\description{
Print a 'lsbclust' object.
}
