% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.lsbclust.R
\name{sim.lsbclust}
\alias{sim.lsbclust}
\title{Simulate from an LSBLCUST model}
\usage{
sim.lsbclust(N = 100, nclust = 5, J = 10, K = 8, ndim = 2, sd = 1)
}
\arguments{
\item{N}{Integer giving the number of observations required.}

\item{nclust}{Integer giving the number of clusters required.}

\item{J}{Integer giving the number of rows to sample.}

\item{K}{Integer giving the number of columns ot sample.}

\item{ndim}{Integer giving the desired dimensionality of the sampled cluster means.}

\item{sd}{The sd argument of \code{\link{rnorm}}.}
}
\description{
This is a simple functiomn to simulate data from an LSBCLUST model.
}

