% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsv.R
\name{simsv}
\alias{simsv}
\title{Randomly Generate Positive Singular Values}
\usage{
simsv(nclust, ndim = 2, mins = 1, max = 5)
}
\arguments{
\item{nclust}{Integer giving the number of clusters for which to sample singular values.}

\item{ndim}{Integer; the number of singular values required.}

\item{mins}{Numeric vector of length \code{ndim} giving the minimum values for the 
respective singular values.}

\item{max}{Numeric value giving the maximum possible value for the mean of the cluster-
specific singular value distribution, relative to the \code{mins}}
}
\description{
Generate random singular values for a specified number of clusters
for use in simulations. A mixture distribution is used with truncation to ensure that the
singular values differ between clusters, are ordered, and are nonnegative.
}
