% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.int.lsbclust.R
\name{summary.int.lsbclust}
\alias{summary.int.lsbclust}
\title{Summary Method for Class "int.lsbclust"}
\usage{
\method{summary}{int.lsbclust}(object, digits = 3, ...)
}
\arguments{
\item{object}{An object of class 'int.lsbclust'.}

\item{digits}{The number of digits in the printed output.}

\item{\dots}{Unimplemented.}
}
\description{
Some goodness-of-fit diagnostics are provided for all three margins.
}
