% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lsbclust.R
\name{summary.lsbclust}
\alias{summary.lsbclust}
\title{Summary Method for Class "lsbclust"}
\usage{
\method{summary}{lsbclust}(object, digits = 3, ...)
}
\arguments{
\item{object}{An object of class 'lsbclust'.}

\item{digits}{The number of digits in the printed output.}

\item{\dots}{Unimplemented.}
}
\description{
Summarize a \code{lsbclust} object.
}
