% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluedefinitions.R
\name{cl_class_ids.int.lsbclust}
\alias{cl_class_ids.int.lsbclust}
\alias{is.cl_partition.int.lsbclust}
\alias{is.cl_hard_partition.int.lsbclust}
\alias{cl_class_ids.lsbclust_sim_part}
\alias{is.cl_partition.lsbclust_sim_part}
\alias{is.cl_hard_partition.lsbclust_sim_part}
\alias{cl_class_ids.T3Clusf}
\alias{is.cl_partition.T3Clusf}
\alias{is.cl_hard_partition.T3Clusf}
\alias{cl_class_ids.akmeans}
\alias{is.cl_partition.akmeans}
\alias{is.cl_hard_partition.akmeans}
\title{S3 export}
\usage{
\method{cl_class_ids}{int.lsbclust}(x)

\method{is.cl_partition}{int.lsbclust}(x)

\method{is.cl_hard_partition}{int.lsbclust}(x)

\method{cl_class_ids}{lsbclust_sim_part}(x)

\method{is.cl_partition}{lsbclust_sim_part}(x)

\method{is.cl_hard_partition}{lsbclust_sim_part}(x)

\method{cl_class_ids}{T3Clusf}(x)

\method{is.cl_partition}{T3Clusf}(x)

\method{is.cl_hard_partition}{T3Clusf}(x)

\method{cl_class_ids}{akmeans}(x)

\method{is.cl_partition}{akmeans}(x)

\method{is.cl_hard_partition}{akmeans}(x)
}
\arguments{
\item{x}{An object of class \code{int.lsclust}}
}
\description{
These export
into the framework set out in package \pkg{clue}.
}
